requirements.txt
allauth
certifi-2018.4.16
chardet-3.0.4
coreapi-2.3.3
coreschema-0.0.4
defusedxml-0.5.0
django-2.0.4
django-allauth-0.35.0
django-cors-headers-2.2.0
django-rest-framework-0.1.0
django-rest-auth-0.9.3
djangorestframework-3.8.2
dj-database-url-0.5.0
drf-extensions-0.3.1
idna-2.6
itypes-1.1.0
jinja2-2.10
MarkupSafe-1.0
oauthlib-2.0.7
psycopg2-2.7.4
python3-openid-3.1.0
pytz-2018.4
python-dotenv-0.8.2
raven==6.7.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
uritemplate==3.0.0